+++ Fluent.Ribbon/Controls/ComboBox.cs
@@ -51,6 +51,22 @@ public class ComboBox : System.Windows.Controls.ComboBox, IQuickAccessItemProvid
 
     #region Properties
 
+    #region CornerRadius
+
+    /// <summary>
+    /// Gets or sets the degree to which the corners of the control are rounded.
+    /// </summary>
+    public CornerRadius CornerRadius
+    {
+        get => (CornerRadius)this.GetValue(CornerRadiusProperty);
+        set => this.SetValue(CornerRadiusProperty, value);
+    }
+
+    /// <summary>Identifies the <see cref="CornerRadius"/> dependency property.</summary>
+    public static readonly DependencyProperty CornerRadiusProperty = Border.CornerRadiusProperty.AddOwner(typeof(ComboBox));
+
+    #endregion
+
     #region Size
 
     /// <inheritdoc />