+++ Fluent.Ribbon/Controls/InRibbonGallery.cs
@@ -73,6 +73,22 @@ public class InRibbonGallery : Selector, IScalableRibbonControl, IDropDownContro
 
     #region Properties
 
+    #region CornerRadius
+
+    /// <summary>
+    /// Gets or sets the degree to which the corners of the control are rounded.
+    /// </summary>
+    public CornerRadius CornerRadius
+    {
+        get => (CornerRadius)this.GetValue(CornerRadiusProperty);
+        set => this.SetValue(CornerRadiusProperty, value);
+    }
+
+    /// <summary>Identifies the <see cref="CornerRadius"/> dependency property.</summary>
+    public static readonly DependencyProperty CornerRadiusProperty = Border.CornerRadiusProperty.AddOwner(typeof(InRibbonGallery));
+
+    #endregion
+
     #region Size
 
     /// <inheritdoc />