+++ Fluent.Ribbon/Controls/RibbonGroupBox.cs
@@ -63,6 +63,22 @@ public class RibbonGroupBox : HeaderedItemsControl, IQuickAccessItemProvider, ID
     /// </summary>
     public ContentControl? CollapsedHeaderContentControl { get; private set; }
 
+    #region CornerRadius
+
+    /// <summary>
+    /// Gets or sets the degree to which the corners of the control are rounded.
+    /// </summary>
+    public CornerRadius CornerRadius
+    {
+        get => (CornerRadius)this.GetValue(CornerRadiusProperty);
+        set => this.SetValue(CornerRadiusProperty, value);
+    }
+
+    /// <summary>Identifies the <see cref="CornerRadius"/> dependency property.</summary>
+    public static readonly DependencyProperty CornerRadiusProperty = Border.CornerRadiusProperty.AddOwner(typeof(RibbonGroupBox));
+
+    #endregion
+
     #region KeyTip
 
     /// <inheritdoc />